+++ include/core/math/vector/vector2i.hpp
@@ -10,10 +10,10 @@ namespace core::math {
 #pragma warning(disable: 4201)
 	struct Vector2i {
 		union {
-			struct { s64 x, y; };
-			struct { s64 w, h; };
-			struct { s64 min, max; };
-			s64 data[2];
+			struct { s32 x, y; };
+			struct { s32 w, h; };
+			struct { s32 min, max; };
+			s32 data[2];
 		};
 	};
 #pragma warning(pop)
@@ -30,8 +30,8 @@ namespace core::math {
 	constexpr Vector2i operator-(Vector2i v1, Vector2i v2) { return { (v1.x - v2.x), (v1.y - v2.y) }; }
 	constexpr Vector2i operator*(Vector2i v1, Vector2i v2) { return { (v1.x * v2.x), (v1.y * v2.y) }; }
 
-	constexpr Vector2i operator*(Vector2i v, s64 scaler) { return { (v.x * scaler), (v.y * scaler) }; }
-	constexpr Vector2i operator*(s64 scaler, Vector2i v) { return v * scaler; }
+	constexpr Vector2i operator*(Vector2i v, s32 scaler) { return { (v.x * scaler), (v.y * scaler) }; }
+	constexpr Vector2i operator*(s32 scaler, Vector2i v) { return v * scaler; }
 
 	constexpr bool operator==(Vector2i v1, Vector2i v2) { return (v1.x == v2.x) && (v1.y == v2.y); }
 	constexpr bool operator!=(Vector2i v1, Vector2i v2) { return (v1.x != v2.x) || (v1.y != v2.y); }
