+++ src/core/math/vector/vector2.cpp
@@ -5,11 +5,11 @@
 namespace core::math::vector {
 
 	// functions
-	f64 length(Vector2 v) { return sqrt((v.x * v.x) + (v.y * v.y)); }
+	f32 length(Vector2 v) { return sqrt((v.x * v.x) + (v.y * v.y)); }
 
 	Vector2 normalize(Vector2 v) {
-		f64 magnitude = length(v);
-		ASSERT(magnitude != 0.0);
+		f32 magnitude = length(v);
+		ASSERT(magnitude != 0.0f);
 		return { (v.x / magnitude), (v.y / magnitude) };
 	}
