+++ src/core/math/vector/vector3.cpp
@@ -5,11 +5,11 @@
 namespace core::math::vector {
 
 	// functions
-	f64 length(Vector3 v) { return sqrt((v.x * v.x) + (v.y * v.y) + (v.z * v.z)); }
+	f32 length(Vector3 v) { return sqrt((v.x * v.x) + (v.y * v.y) + (v.z * v.z)); }
 
 	Vector3 normalize(Vector3 v) {
-		f64 magnitude = length(v);
-		ASSERT(magnitude != 0.0);
+		f32 magnitude = length(v);
+		ASSERT(magnitude != 0.0f);
 		return { (v.x / magnitude), (v.y / magnitude), (v.z / magnitude) };
 	}
